third_party/blink/web_tests/fast/dom/shadow/sibling-rules-dynamic-changes-expected.txt

Test dynamically changing sibling rule matches in ShadowRoots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Dynamically change the :first-child
PASS getComputedStyle(first).color is "rgb(255, 0, 0)"
PASS getComputedStyle(second).color is "rgb(0, 0, 0)"
Dynamically change the :last-child
PASS getComputedStyle(last).color is "rgb(255, 0, 0)"
PASS getComputedStyle(secondToLast).color is "rgb(0, 0, 0)"
Dynamically change what a + combinator matches
PASS getComputedStyle(span).color is "rgb(0, 0, 0)"
Dynamically change what a ~ combinator matches
PASS getComputedStyle(span).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
